Make toggle more accessible
Make the New Arrivals/On Order toggle more accessible by:

1. Making the checkboxes visually hidden rather than completely hidden
2. Making the container of the checkboxes a fieldset rather than a div
3. Providing a legend for the fieldset (and visually hiding it)
4. Overriding the default fieldset padding to keep the visual design as
expected.
5. Adding tabindex attributes where needed.

With these changes you can tab to the fieldset and then tab to the
checkboxes. Once the checkboxes are active you can use the arrow keys to
switch back and forth. An enhancement would be to add a visual emphasis
to the labels when the corresponding checkboxes are active. Currently,
because the checkboxes are visually hidden, it looks like the keyboard
focus disappears.

With VoiceOver on, the fieldset legend is read and the checkoxes are
announced.