Make selected filters keyboard accessible
Wrap selected filters in <a> tags so that can be focussed and
deactivated by keyboard. Add CSS to keep them styled properly.

This is a minimal solution for keyboard access only. A new page load is
prevented by setting the link href to "#", but this might cause the page
to jump, and it also might interfere with any history manipulations of
the URL we are doing.

Screen reader accessibility is improved slightly since the filter can
now be read, but more work needs to be done to provide better screen
reader context and grouping.